Expand MotorControl with movement methods & member variables
Speed is saved in the MotorControl class, as well as a number of static speed variables that can be reused throughout the project as default speeds. The class also has a modifier for each wheel we can use in case the motors aren't properly calibrated / we run into issues due to different weight / grip distributions.